# varint: improve encode/decode perf