@bitspill 2023-10-02
I misread the documentation on checked_shl it's validating the argument not the operation and was a stale test result I looked at before pushing.

Now explicitly checking the top 7 bits before performing the shift and erroring as required