@casey 2023-10-06
Thanks for the PR! The canonical implementation is now being added to ord, check out [this PR](https://github.com/ordinals/ord/pull/2491).

The varint implementation now writes varints into a `Vec<u8>` so it doesn't allocate a separate vec.